gitsign: fix for RejectExpiredKeys logic
Current RejectExpiredKeys didn't work as CheckDetachedSignature
validates expiration date internally.

It looks like go-crypto library does not allow signature checks
with completely ignoring exipration times (as Github UI allows
for example). With this change we allow (option for) keys to be expired,
but the signature creation time can not be after key expiry.

If we would set the reference time to the key creation time
that would cause a different error where validation would not work
because signatures are created in future.

Signed-off-by: Tonis Tiigi <tonistiigi@gmail.com>